+++ DSP_Plugins.GalacticScale/Scripts/GalacticScale2.0/Patches/PlanetRawData/CalcVerts.cs
@@ -3,17 +3,14 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace GalacticScale
-{
-    public partial class PatchOnPlanetRawData
-    {   public static Dictionary<int, Vector3[]> verDict = new Dictionary<int, Vector3[]>();
+namespace GalacticScale {
+    public partial class PatchOnPlanetRawData {
+        public static Dictionary<int, Vector3[]> verDict = new Dictionary<int, Vector3[]>();
         public static Dictionary<int, int[]> inDict = new Dictionary<int, int[]>();
-        
+
         [HarmonyPrefix, HarmonyPatch(typeof(PlanetRawData), "CalcVerts")]
-        public static bool CalcVerts(ref PlanetRawData __instance, ref int[] ___indexMap80, ref int[] ___indexMap200, ref Vector3[] ___verts80, ref Vector3[] ___verts200)
-        {
-            if (verDict.ContainsKey(__instance.precision))
-            {
+        public static bool CalcVerts(ref PlanetRawData __instance, ref int[] ___indexMap80, ref int[] ___indexMap200, ref Vector3[] ___verts80, ref Vector3[] ___verts200) {
+            if (verDict.ContainsKey(__instance.precision)) {
                 Array.Copy(verDict[__instance.precision], __instance.vertices, verDict[__instance.precision].Length);
                 Array.Copy(inDict[__instance.precision], __instance.indexMap, inDict[__instance.precision].Length);
                 return false;
@@ -25,8 +22,7 @@ namespace GalacticScale
             }
             int stride = (__instance.precision + 1) * 2;
             int substride = __instance.precision + 1;
-            for (int j = 0; j < __instance.dataLength; j++)
-            {
+            for (int j = 0; j < __instance.dataLength; j++) {
                 int num3 = j % stride;
                 int num4 = j / stride;
                 int num5 = num3 % substride;
@@ -34,15 +30,14 @@ namespace GalacticScale
                 int num7 = (((num3 >= substride) ? 1 : 0) + ((num4 >= substride) ? 1 : 0) * 2) * 2 + ((num5 < num6) ? 1 : 0); //
                 float num8 = ((num5 < num6) ? num5 : (__instance.precision - num5));
                 float num9 = ((num5 < num6) ? (__instance.precision - num6) : num6);
-                float num10 = (float)__instance.precision - num9;
-                num9 /= (float)__instance.precision;
+                float num10 = __instance.precision - num9;
+                num9 /= __instance.precision;
                 num8 = ((!(num10 > 0f)) ? 0f : (num8 / num10));
                 int num11 = 0;
                 Vector3 a;
                 Vector3 a2;
                 Vector3 b;
-                switch (num7)
-                {
+                switch (num7) {
                     case 0:
                         a = PlanetRawData.poles[2];
                         a2 = PlanetRawData.poles[0];
@@ -101,22 +96,18 @@ namespace GalacticScale
                 ref Vector3 reference = ref __instance.vertices[j];
                 reference = Vector3.Slerp(Vector3.Slerp(a, b, num9), Vector3.Slerp(a2, b, num9), num8);
                 int num12 = __instance.PositionHash(__instance.vertices[j], num11);
-                if (__instance.indexMap[num12] == -1)
-                {
+                if (__instance.indexMap[num12] == -1) {
                     __instance.indexMap[num12] = j;
                 }
             }
             int num13 = 0;
-            for (int k = 1; k < __instance.indexMapDataLength; k++)
-            {
-                if (__instance.indexMap[k] == -1)
-                {
+            for (int k = 1; k < __instance.indexMapDataLength; k++) {
+                if (__instance.indexMap[k] == -1) {
                     __instance.indexMap[k] = __instance.indexMap[k - 1];
                     num13++;
                 }
             }
-            if (!verDict.ContainsKey(__instance.precision))
-            {
+            if (!verDict.ContainsKey(__instance.precision)) {
                 int precision = __instance.precision;
                 int vertLength = __instance.vertices.Length;
                 int indexMapLength = __instance.indexMap.Length;
