+++ DSP_Plugins.GalacticScale/Scripts/GalacticScale2.0/Patches/UIPlanetDetail/SetResCount.cs
@@ -1,16 +1,14 @@
 ﻿using HarmonyLib;
 using UnityEngine;
 
-namespace GalacticScale
-{
+namespace GalacticScale {
 
-    public partial class PatchOnUIPlanetDetail
-    {
-        [HarmonyPrefix, HarmonyPatch(typeof(UIPlanetDetail),"SetResCount")]
+    public partial class PatchOnUIPlanetDetail {
+        [HarmonyPrefix, HarmonyPatch(typeof(UIPlanetDetail), "SetResCount")]
         public static bool SetResCount(int count, ref RectTransform ___rectTrans, ref RectTransform ___paramGroup) // Adjust the height of the PlanetDetail UI to allow for Radius Text
         {
-            ___rectTrans.sizeDelta = new Vector2(___rectTrans.sizeDelta.x, (float)(190 + count * 20) + 20f);
-            ___paramGroup.anchoredPosition = new Vector2(___paramGroup.anchoredPosition.x, (float)(-90 - count * 20));
+            ___rectTrans.sizeDelta = new Vector2(___rectTrans.sizeDelta.x, 190 + count * 20 + 20f);
+            ___paramGroup.anchoredPosition = new Vector2(___paramGroup.anchoredPosition.x, -90 - count * 20);
             return false;
         }
